Refactor for mix credo
